requirements.txt
fastapi
uvicorn
ultralytics
Pillow
numpy
python-multipart
requests